NextStateAssigner: refactored to the state package
